add privacy switch support for smt-client
Detect if smt-client runs on SMT server and check if forwardRegistration is allowed.
If not, smt-client is not allowed to run on the SMT server as it would contact NCC

With the new feature of NCC to also send jobs via the SMT JobQueue into a customers SMT server
he opens his JobQueue to NCC which in turn can gather information about the customers machines.

The new feature is about managing, maintaining these machines in the NCC WebUI incl. inventory.
Thus new jobs were added to smt-client and SMT server. To enable this feature a customer just
needs to install the smt-client package on the SMT server.