Branch Repair: pull all changes to smt-client in master to branch_SMT11
A smt-client package was incorreclty released _for_ SLE11 _from_ the master branch as smt-client 0.0.13.
The changes did not went into branch_SMT11 and its version number remained at 0.0.10 until a bugfix of bug#660191 was made which led to version 0.0.11.
This version never existed: it was not accepted by autobuild as the changes were not fast-forward and the version number lower than the released one.

With this commit the smt-client package is at the current release state in branch_SMT11. Please ONLY make updates for SLE11 in this branch.
Cherry-pick every bugfix into master, but do not release SLE11 updates from master.