set guid if missing in job xml, fix for (bnc#723583)
When old smt clients are reporting their results to the SMT server
they do not include their GUID in the job XML (until verion 0.0.14).
The guid is now sanity checked by the SMT server to prevent updates
of foreign jobs. However the old style needs to be supported until
all smt clients are updated, thus the client guid is just set and
not verified if missing.

Note: in a later release we will have to deprecate the old style and
mark the client status to help admins find old smt clients.
The new style is (only) needed for use with the new job types that
send data upstream to a cascaded SMT server (see as well details
about the inventory, report and createjob job types).