introduce new parallel XML format in smt-job
To reliably transmit data to the smt-client it was necessary to introduce
a more strict XML format. As the previously used perl module XML-Simple
mangles XML data in a way that input and output differed it had to be replaced
The more strict XML-format provides a new way to transmit options for jobs.
The options used to reside as attributes in the arguments element. As attributes
no CDATA or lists of values are possible. Thus an 'options' element is introduced.

Note:
For legacy jobs (types 1-8) all options are present as attributes in the arguments
tag as well as in the new options element inside arguments:

<job type="softwarepush" name="install whois" verbose="0">
<arguments agreelicenses="1" force="1">
<options>
<agreelicenses>1</agreelicenses>
<force>1</force>
</options>
<packages>
<package>whois</package>
</packages>
</arguments>
</job>

The old format must be supported (as far as logically possible) to be backward-compatible
with old smt clients.

The new jobs 51-53 do not use attributes at all, so they do not have the options element.