custom output for update job (otherwise exeeds database limit)
The fields for STDOUT and STDERR are BLOBS in the MySQL database,
thus limiting input to 64kB. The new approach to count patches
produces more than 1MB of output. This renders the usage of this
output in a verbosejob mode useless. So the patchstatus job now
creates a small custom summary on stdout in verbosejob mode.

Later:
In a later version we might want to migrate the BLOB fields to LONGBLOB.