fix proxy configuration in smt-client (bnc#777755)
Drop the call to no_proxy as it is no longer needed as workaround for SSLeay
as we no longer use SSLeay in smt-client (was dropped in an early developer version).

The call to env_proxy was missing in order to set the proxy settings
that have been written to the respective environment variables.

Note:
The smt-client reads its proxy settings directly from /etc/sysconfig/smt-client
If no proxy is set there, the values already present in the environment will be used.
So the proxy settings from /etc/sysconfig/proxy will only come into play when smt-agent
is called from an interactive shell which sources the proxy variables via the profile.