make files with wrong checksum as invalid and remove them before mirror again
FIXES:
broken metadata on the server. New repomd.xml with new checksum for
primary.xml, but primary.xml is still the old one.
The outdated() function in Job class return false in such a case and
the file is stated as up-to-date.
We want and error in this case, because the checksums do not match.